ACSL 2022-2023 - Contest 1 - Next Base - Junior Division

+++ Java/Contest 1/2022 - 2023/NextBaseJunior.java
@@ -29,7 +29,7 @@ public class NextBaseJunior {
             int sum = 0;
 
             while (number > 0) {
-                sum += number % base;
+                sum += (int)(number % base);
                 number /= base;
             }
